Time: 33 ms (20.28%), Space: 16.2 MB (99.99%) - LeetHub

+++ 0018-4sum/0018-4sum.cpp
@@ -13,7 +13,7 @@ public:
             
             while(k<l) {
             
-                int sum = nums[i]+nums[j];
+                long long sum = nums[i]+nums[j];
                 sum+=nums[k]+nums[l];
                 
                 if(sum == target) {
